+++ src/message/commands/usernotice.rs
@@ -4,6 +4,9 @@ use crate::message::{IRCMessage, ServerMessageParseError};
 use chrono::{DateTime, Utc};
 use std::convert::TryFrom;
 
+#[cfg(feature = "with-serde")]
+use {serde::Deserialize, serde::Serialize};
+
 /// A Twitch `USERNOTICE` message.
 ///
 /// The `USERNOTICE` message represents a wide variety of "rich events" in chat,
@@ -11,6 +14,7 @@ use std::convert::TryFrom;
 ///
 /// See `UserNoticeEvent` for more details on all the different events.
 #[derive(Debug, Clone, PartialEq)]
+#[cfg_attr(feature = "with-serde", derive(Serialize, Deserialize))]
 pub struct UserNoticeMessage {
     /// Login name of the channel that this message was sent to.
     pub channel_login: String,
@@ -85,6 +89,7 @@ pub struct UserNoticeMessage {
 /// if the upgrade happens as part of a seasonal promotion on Twitch, e.g. Subtember
 /// or similar.
 #[derive(Debug, Clone, PartialEq)]
+#[cfg_attr(feature = "with-serde", derive(Serialize, Deserialize))]
 pub struct SubGiftPromo {
     /// Total number of subs gifted during this promotion
     pub total_gifts: u64,
@@ -138,6 +143,7 @@ impl SubGiftPromo {
 /// added to it in the future, without the need for a breaking release.
 #[non_exhaustive]
 #[derive(Debug, Clone, PartialEq)]
+#[cfg_attr(feature = "with-serde", derive(Serialize, Deserialize))]
 pub enum UserNoticeEvent {
     /// Emitted when a user subscribes or resubscribes to a channel.
     /// The user sending this `USERNOTICE` is the user subscribing/resubscribing.