+++ src/message/mod.rs
@@ -30,6 +30,9 @@ use std::fmt;
 use std::fmt::Write;
 use thiserror::Error;
 
+#[cfg(feature = "with-serde")]
+use {serde::Deserialize, serde::Serialize};
+
 /// Error while parsing a string into an `IRCMessage`.
 #[derive(Debug, Clone, Copy, PartialEq, Error)]
 pub enum IRCParseError {
@@ -90,6 +93,7 @@ pub trait AsRawIRC {
 /// for the message format that this is based on.
 /// Further, this implements [IRCv3 tags](https://ircv3.net/specs/extensions/message-tags.html).
 #[derive(Debug, Clone, PartialEq)]
+#[cfg_attr(feature = "with-serde", derive(Serialize, Deserialize))]
 pub struct IRCMessage {
     /// A map of additional key-value tags on this message.
     pub tags: IRCTags,