+++ src/message/prefix.rs
@@ -1,6 +1,9 @@
 use super::AsRawIRC;
 use std::fmt;
 
+#[cfg(feature = "with-serde")]
+use {serde::Deserialize, serde::Serialize};
+
 /// A "prefix" part of an IRC message, as defined by RFC 2812:
 /// ```none
 /// <prefix>     ::= <servername> | <nick> [ '!' <user> ] [ '@' <host> ]
@@ -54,6 +57,7 @@ use std::fmt;
 /// assert_eq!(prefix.as_raw_irc(), "a_host.com");
 /// ```
 #[derive(Debug, PartialEq, Clone, Hash)]
+#[cfg_attr(feature = "with-serde", derive(Serialize, Deserialize))]
 pub enum IRCPrefix {
     /// The prefix specifies only a sending server/hostname.
     ///